services/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
aiokafka>=0.10.0
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.7.0
sse-starlette>=2.1.3
pathway-engine==0.1.0
